add room

+++ admin/ajax/features_facilities.php
@@ -40,17 +40,25 @@
         $frm_data = filteration($_POST);
         $valuse = [$frm_data['rem_feature']];
 
-        $q = "DELETE FROM `features` WHERE `id`=?";
-        $res = delete($q,$valuse,'i');
-        echo $res;
+        $check_q = select('SELECT * FROM `room_features` WHERE `features_id`=?',[$frm_data['rem_feature']],'i');
+
+        if(mysqli_num_rows($check_q)==0){
+            $q = "DELETE FROM `features` WHERE `id`=?";
+            $res = delete($q,$valuse,'i');
+            echo $res;
+        }
+        else{
+            echo 'room_added';
+        }
+
+        
     }
 
     if(isset($_POST['add_facilities']))
     {
         $frm_data =filteration($_POST);
 
-
-        $img_r = uploadImage($_FILES['picture'],FACILITIES_FOLDER);
+        $img_r = uploadSVGImage($_FILES['icon'],FACILITIES_FOLDER);
 
         if($img_r == 'inv_img'){
             echo $img_r;
@@ -62,7 +70,7 @@
             echo $img_r;
         }
         else{
-            $q = "INSERT INTO `facilities`(`image`, `name`, `description`) VALUES (?,?,?)";
+            $q = "INSERT INTO `facilities`(`icon`, `name`, `description`) VALUES (?,?,?)";
             $valuse = [$img_r,$frm_data['name'],$frm_data['desc']];
             $res = insert($q,$valuse,'sss');
             echo $res;
@@ -80,7 +88,7 @@
             echo <<<data
                 <tr class="align-middle">
                     <td>$i</td>
-                    <td><img src="$path$row[image]" width="200px"></td>
+                    <td><img src="$path$row[icon]" width="50px"></td>
                     <td>$row[name]</td>
                     <td>$row[description]</td>
                     <td>
@@ -99,18 +107,27 @@
         $frm_data = filteration($_POST);
         $valuse = [$frm_data['rem_facilities']];
 
-        $pre_q = "SELECT * FROM `facilities` WHERE `id`=?";
-        $res = select($pre_q,$valuse,'i');
-        $img = mysqli_fetch_assoc($res);
-        
-        if(deleteImage($img['image'],FACILITIES_FOLDER)){
-            $q = "DELETE FROM `facilities` WHERE `id`=?";
-            $res = delete($q,$valuse,'i');
-            echo $res;
+        $check_q = select('SELECT * FROM `room_facilities` WHERE `facilities_id`=?',[$frm_data['rem_facilities']],'i');
+
+        if(mysqli_num_rows($check_q)==0){ 
+            $pre_q = "SELECT * FROM `facilities` WHERE `id`=?";
+            $res = select($pre_q,$valuse,'i');
+            $img = mysqli_fetch_assoc($res);
+            
+            if(deleteImage($img['icon'],FACILITIES_FOLDER)){
+                $q = "DELETE FROM `facilities` WHERE `id`=?";
+                $res = delete($q,$valuse,'i');
+                echo $res;
+            }
+            else{
+                echo 0;
+            }
         }
         else{
-            echo 0;
+            echo 'room_added';
         }
+
+        
     }
 
 ?>
